split out queries into separate files
