+++ utils/queries/alerts.ts
@@ -1,0 +1,15 @@
+import gql from "graphql-tag";
+
+export const ALL_ALERTS = gql`
+  query ALL_ALERTS {
+    alerts {
+      id
+      guid
+      title
+      description
+      valid_till
+      updated_at
+      color
+    }
+  }
+`;
