+++ utils/queries/index.ts
@@ -1,0 +1,19 @@
+import { ARTICLES_QUERY } from "./articles";
+import { MODELS_QUERY, MODELS_QUERY_BRIEF } from "./models";
+import { ALL_SCALES_QUERY } from "./scales";
+import { ALL_MANUFACTURERS_QUERY } from "./manufacturers";
+import { ALL_ALERTS } from "./alerts";
+import { GET_ALL_GALLERIES, GET_ALL_GALLERIES_BRIEF } from "./galleries";
+import { ALL_TAGS } from "./tags";
+
+export {
+  ARTICLES_QUERY,
+  MODELS_QUERY,
+  MODELS_QUERY_BRIEF,
+  ALL_ALERTS,
+  ALL_MANUFACTURERS_QUERY,
+  ALL_SCALES_QUERY,
+  GET_ALL_GALLERIES,
+  GET_ALL_GALLERIES_BRIEF,
+  ALL_TAGS,
+};
