+++ utils/queries/tags.ts
@@ -1,0 +1,11 @@
+import gql from "graphql-tag";
+
+export const ALL_TAGS = gql`
+  query ALL_TAGS {
+    tags {
+      id
+      title
+      slug
+    }
+  }
+`;
